Fix more checkboxes of other QT types
There are a few other checkbox types which were not showing up. I copy
these from a different theme to try and ensure full coverage of the
different checkbox types.

Unchecked box works ok, you see a black box for all the below types, so
that is why I am not adding this for the unchecked version.